Move warning suppresion to environments (example)
This shows how to have environments indicate which warnings should be ignored.
The current apporach has this information in the test suite, which won't
work for a new custom environment.

Two environments demonstrate use of the new system:

* chess
* aec_rps (in the code examples)

There is no change to the actual behaviour of either environment, it is
only a change to the metadata. The remaining petting zoo envs would also
need the updated metadata to move completely to the new system.
See pettingzoo/test/_trait_flags.py for more details

This behaviour is completely optional.
All custom environments should work as-is with no changes required.