Move warning suppresion to environments (example, version 3)
This shows how to have environments indicate which warnings should be ignored.
The current apporach has this information in the test suite, which won't
work for a new custom environment.

Two environments demonstrate use of the new system:
chess
aec_rps (in the code examples)

There is no change to the actual behaviour of either environment, it is
an additional, optional function. The remaining petting zoo envs would
also need the added function to move completely to the new system.
See pettingzoo/test/pz_warnings.py for more details

This behaviour is completely optional.
All custom environments should work with no changes required.