Fix value of FPS in KAZ
This was being used as both the render speed and a scaling factor for
movement speed. The problem with this is that it slowed movement speed
but not action rate. So while doubling the FPS might still yield the
same absolute total movement, it led to twice as many action so twice
as many chances to attack, twice as many reward chances, etc

The FPS was not a user config option anyway, so removing it should not
impact the env behavior unless someone edited the env locally.

FPS has been removed and the render_fps is hard coded at 15 (the
previous value). Object speeds no longer depend on FPS.